+++ Sources/HabitatCore/PolicyReasonCatalog+CommandFamilies.swift
@@ -5,6 +5,5 @@ extension PolicyReasonCatalog {
     ]
 
     static let catalogCommandFamilies = dynamicAskFirstCommandFamilies
-        + baselineAskFirstCommandFamilies
-        + baselineForbiddenCommandFamilies
+        + baselineCommandFamilies
 }